pyproject.toml
[project]
name = "ppt-yahe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["python-pptx>=1.0", "pillow>=12.2"]
